Stop using dict as default value for func arg
This gets you into trouble in python, because the same reference to the
same actual object gets used every time the function is called. So if
there's a dict there, and you modify it, your modifications will be in
the default value the next time you call the function.

TLDR: Only use immutable types for default values.